GnssLocationProvider to remember the Networks and connection types
When receiving connection lost callback, the network object passed
along can no longer be mapped to an active NetworkInfo object. So
GnssLocationProvider is not able to find out whether that was a wifi
or wwan connection. When both are connected and later one disconnected,
it is not able to tell which one of the two was dropped.

Bug: 93175534
CRs-Fixed: 2252529

Change-Id: Ibfb1bf548b352d08f7110551952817a17db94ba6
Signed-off-by: DennySPB <dennyspb@gmail.com>
Signed-off-by: VenkatVishalV <venkatvishal124@gmail.com>